Add support for versioned ChainObjects
